+++ afat/__init__.py
@@ -4,5 +4,5 @@ app config
 
 default_app_config: str = "afat.apps.AfatConfig"
 
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 __title__ = "Fleet Activity Tracking"